+++ pkg/analyzer_plugin/lib/utilities/analyzer_converter.dart
@@ -61,7 +61,7 @@ class AnalyzerConverter {
         error.message,
         errorCode.name.toLowerCase(),
         contextMessages: contextMessages,
-        correction: error.correction,
+        correction: error.correctionMessage,
         hasFix: true);
   }